Refactor prioritization discovery package
