+++ README.md
@@ -19,11 +19,11 @@ You can browse compiled releases in the [Releases](https://github.com/AutomatedP
 
 ## Description
 
-The `pix_framework.discovery` packge besides root modules contains the following subpackages with additional information located in their README files:
+The `pix_framework.discovery` package besides root modules contains the following subpackages with additional information located in their README files:
 
 - [batch_processing_discovery](src/pix_framework/discovery/batch_processing/)
-- [prioritization_discovery](./src/pix_framework/discovery/prioritization_discovery/)
-- [start_time_estimator](./src/pix_framework/discovery/start_time_estimator/)
+- [prioritization_discovery](src/pix_framework/discovery/prioritization/)
+- [case_attribute_discovery](src/pix_framework/discovery/case_attribute/)
 
 ## Development
